+++ Angular-Front-End/app-front-end/src/app/models/login.model.ts
@@ -2,6 +2,6 @@ import { Identifiable } from './identifiable';
 import { UserModel } from './user.model';
 
 export interface LoginModel extends Identifiable{
-    user: UserModel;
-    msg: string;
+    first: string;
+    second: UserModel;
 }