+++ Services/EventService/src/main/java/io/App/EventService/EventComponent/Event.java
@@ -9,7 +9,8 @@ public class Event {
 	private int cID;
 	private String cName;
 
-	public Event(int id, String name, String start, String end, int cID, String cName) {
+	public Event(int id, String name, String start, String end, int cID,
+			String cName) {
 		this.id = id;
 		this.name = name;
 		this.start = start;
@@ -25,7 +26,7 @@ public class Event {
 	public int getId() {
 		return id;
 	}
-	
+
 	public String getName() {
 		return name;
 	}