+++ Services/EventService/src/main/java/io/App/EventService/EventComponent/EventCatalog.java
@@ -6,6 +6,7 @@ import io.App.EventService.databaseConnection.EventDatabaseConnection;
 import io.App.EventService.dto.EventListWrapper;
 import io.App.EventService.exceptions.EventAlreadyExistsException;
 import io.App.EventService.exceptions.EventDoesNotExistException;
+import io.App.EventService.exceptions.InternalAppException;
 import io.App.EventService.exceptions.NoEventsFromCommunityException;
 
 @SpringBootApplication
@@ -21,23 +22,22 @@ public class EventCatalog {
 		return this.eDC.getAllEvents();
 	}
 
-	public EventListWrapper getEventsFromCommunity(int cID) throws NoEventsFromCommunityException {
+	public EventListWrapper getEventsFromCommunity(int cID)
+			throws NoEventsFromCommunityException {
 		EventListWrapper eLW = this.eDC.getEventsFromCommunity(cID);
-		if(eLW == null) {
+		if (eLW == null) {
 			throw new NoEventsFromCommunityException();
 		}
 		return eLW;
 	}
 
-	public void registerNewEvent(Event event) throws EventAlreadyExistsException {
-		if(this.eDC.getEventByName(event.getName()) != null) {
-			throw new EventAlreadyExistsException();
-		}
+	public void registerNewEvent(Event event)
+			throws EventAlreadyExistsException, InternalAppException {
 		this.eDC.registerNewEvent(event);
 	}
 
 	public void deleteEvent(Event event) throws EventDoesNotExistException {
-		if(this.eDC.getEventByName(event.getName()) == null) {
+		if (this.eDC.getEventByName(event.getName()) == null) {
 			throw new EventDoesNotExistException();
 		}
 		eDC.deleteEvent(event);