+++ Services/EventService/src/main/java/io/App/EventService/dto/EventListWrapper.java
@@ -1,25 +1,27 @@
 package io.App.EventService.dto;
 
-import java.util.List;
+import java.util.ArrayList;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 
 import io.App.EventService.EventComponent.Event;
 
+
 public class EventListWrapper {
 
-	@JsonProperty("listEvents")
-	private List<Event> list;
+    @JsonProperty("listEvents")
+    private ArrayList<Event> list;
 
-	public EventListWrapper() {
-		// For REST only
-	}
 
-	public List<Event> getList() {
-		return list;
-	}
+    public EventListWrapper() {
+        //For REST Only
+    }
 
-	public void setList(List<Event> someList) {
-		this.list = someList;
-	}
+    public ArrayList<Event> getList() {
+        return list;
+    }
+
+    public void setList(ArrayList<Event> someList) {
+        this.list = someList;
+    }
 }