+++ Services/UserCatalogService/src/main/java/io/App/UserManagementService/dto/UserLoginModel.java
@@ -1,21 +1,30 @@
 package io.App.UserManagementService.dto;
 
+import java.io.Serializable;
+
 import com.fasterxml.jackson.annotation.JsonProperty;
 
-public class UserLoginModel {
+public class UserLoginModel implements Serializable {
+
+	private static final long serialVersionUID = 1L;
+
 	@JsonProperty("username")
-	private String uName;
+	private String username;
 
 	@JsonProperty("password")
 	private String password;
 
-	public UserLoginModel(String uName, String password) {
-		this.uName = uName;
+	public UserLoginModel(String username, String password) {
+		this.username = username;
 		this.password = password;
 	}
+	
+	public UserLoginModel() {
+		// For REST only
+	}
 
-	public String getuName() {
-		return uName;
+	public String getUsername() {
+		return username;
 	}
 
 	public String getPassword() {