+++ Services/UserCatalogService/src/main/java/io/App/UserManagementService/main/UserCatalogServiceApplication.java
@@ -2,8 +2,10 @@ package io.App.UserManagementService.main;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.context.annotation.ComponentScan;
 
 @SpringBootApplication
+@ComponentScan(basePackages= {"io.App.UserManagementService"})
 public class UserCatalogServiceApplication {
 
 	public static void main(String[] args) {